+++ src/tracker.ts
@@ -28,7 +28,10 @@ export class Tracker extends BaseFileRef {
 
   private getItems() {
     const content = this.readFile();
-    const lines = content.split("\n").filter(Boolean);
+    const lines = content
+      .split("\n")
+      .filter(Boolean)
+      .filter((line) => !line.startsWith("#"));
     const items = lines.map((line) => line.split(" "));
     return items;
   }